progress on refactoring
