+++ Language/Java/Lenses/Lenses.hs
@@ -1,0 +1,11 @@
+module Language.Java.Lenses.Lenses
+    ( module Language.Java.Syntax
+    , module Language.Java.Lenses.Syntax
+    , module Language.Java.Lenses.Syntax.Types
+    --, module Language.Java.Lenses.Syntax.Exp
+) where
+
+import           Language.Java.Lenses.Syntax
+--import           Language.Java.Prisms.Syntax.Exp
+import           Language.Java.Lenses.Syntax.Types
+import           Language.Java.Syntax              ()
