+++ Language/Java/Lenses/Syntax/Exp.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE TemplateHaskell #-}
-module Language.Java.Prisms.Syntax.Exp where
+module Language.Java.Lenses.Syntax.Exp where
 
 import           Language.Java.Prisms.CustomLenses
 import           Language.Java.Syntax.Exp
