--- Language/Java/Prisms/Prisms.hs
@@ -1,11 +1,0 @@
-module Language.Java.Prisms.Prisms
-    ( module Language.Java.Syntax
-    , module Language.Java.Prisms.Syntax
-    , module Language.Java.Prisms.Syntax.Types
-    --, module Language.Java.Prisms.Syntax.Exp
-) where
-
-import           Language.Java.Prisms.Syntax
---import           Language.Java.Prisms.Syntax.Exp
-import           Language.Java.Prisms.Syntax.Types
-import           Language.Java.Syntax              ()
