+++ language-java.cabal
@@ -38,15 +38,15 @@ Library
                         Language.Java.Pretty,
                         Language.Java.Position,
                         Language.Java.Helper,
-                        Language.Java.Prisms.Syntax,
-                        Language.Java.Prisms.Syntax.Exp,
-                        Language.Java.Prisms.Syntax.Types,
-                        Language.Java.Prisms.Prisms,
+                        Language.Java.Lenses.Syntax,
+                        Language.Java.Lenses.Syntax.Exp,
+                        Language.Java.Lenses.Syntax.Types,
+                        Language.Java.Lenses.Lenses,
                         Language.Java.SyntaxClasses
   
   Other-modules:        Language.Java.Syntax.Types,
                         Language.Java.Syntax.Exp,
-                        Language.Java.Prisms.CustomLenses
+                        Language.Java.Lenses.CustomLenses
 
 
 Test-Suite test-java-parse
